+++ pages/top.php
@@ -2,4 +2,5 @@
         <nav class="navbar navbar-expand navbar-light bg-white topbar mb-4 static-top shadow">
 
 </ul>
+<?php checkVer($ver);?>
 </nav>